New email from Simon with new files
From Simon on 23 March 2016:
First, I send you new trees with tip labels that match your sample names. That will facilitate the PGLS analyses. I also send you a new "duplicate.tips" function that also appends a number "1" to the tips and that pads the number with zeros to the left. Again, this increases compatibility with your files.

Now, to properly analyse the bud burst results, I would need the data for each treatment in your "indforGBS.csv" file. This will allow to use proper linear models. The other file, if I understand well, contains marginal means per treatment and is not really appropriate for PGLS (we could try it, but the results would be very crude). So if you can add this info in the "indforGBS" file, I could probably run some analyses rapidly.